Refact: Send 페이지의 Section 컴포넌트들 별도 파일로 분리
