+++ src/components/Send/ImageSection/ImageSection.module.scss
@@ -1,0 +1,32 @@
+.imageSection {
+  width: 100%;
+  display: flex;
+  gap: 5px;
+  padding: 16px 17.5px;
+  overflow-x: auto;
+  border-bottom: 1px solid #8f8f8f;
+
+  > .inputImage {
+    position: relative;
+    > label {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      width: 101px;
+      height: 113px;
+      background-color: gray;
+    }
+    > input {
+      position: absolute;
+      width: 0;
+      height: 0;
+      padding: 0;
+    }
+  }
+
+  > .imagePreview {
+    height: 113px;
+    width: 101px;
+    object-fit: contain;
+  }
+}
