+++ src/components/Send/LocationSection/LocationSection.module.scss
@@ -1,0 +1,44 @@
+.locationSection {
+  display: flex;
+  flex-direction: column;
+  padding: 15.5px 0;
+  border-bottom: 1px solid #8f8f8f;
+
+  > .label {
+    padding: 0 16px;
+    font-size: 16px;
+    font-weight: 400;
+    margin-bottom: 25px;
+  }
+  > .chips {
+    display: flex;
+    padding: 0 16px;
+    margin-bottom: 16px;
+    overflow-x: scroll;
+    flex-wrap: nowrap;
+    gap: 5px;
+
+    > .locationChip {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      flex: 0 0 auto;
+      padding: 0 8px;
+      font-size: 14px;
+      color: #5e5e5e;
+      background: #d9d9d9;
+      height: 29px;
+      border-radius: 8px;
+      white-space: nowrap;
+    }
+  }
+  > .mapContainer {
+    padding: 0 16px;
+    margin-bottom: 28px;
+
+    > .map {
+      width: 100%;
+      height: 156px;
+    }
+  }
+}
