+++ src/components/Send/SubmitSection/SubmitSection.module.scss
@@ -1,0 +1,23 @@
+.submitSection {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 30px 0;
+  > .mainPasteButton {
+    width: 85px;
+    height: 41px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font: inherit;
+    font-size: 16px;
+    font-weight: 700;
+    color: white;
+    border: none;
+    background-color: #d9d9d9;
+    &.active {
+      background-color: #369ed9;
+    }
+  }
+}
