+++ src/components/Send/TextSection/TextSection.module.scss
@@ -1,0 +1,14 @@
+.textSection {
+  width: 100%;
+  border-bottom: 1px solid #8f8f8f;
+
+  > textarea {
+    border: none;
+    font: inherit;
+    font-size: 16px;
+    font-weight: 400;
+    width: 100%;
+    padding: 20px 16.5px;
+    resize: none;
+  }
+}
