+++ src/pages/Send/Send.module.scss
@@ -31,153 +31,3 @@
     }
   }
 }
-
-.locationSection {
-  display: flex;
-  flex-direction: column;
-  padding: 15.5px 0;
-  border-bottom: 1px solid #8f8f8f;
-
-  > .label {
-    padding: 0 16px;
-    font-size: 16px;
-    font-weight: 400;
-    margin-bottom: 25px;
-  }
-  > .chips {
-    display: flex;
-    padding: 0 16px;
-    margin-bottom: 16px;
-    overflow-x: scroll;
-    flex-wrap: nowrap;
-    gap: 5px;
-
-    > .locationChip {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      flex: 0 0 auto;
-      padding: 0 8px;
-      font-size: 14px;
-      color: #5e5e5e;
-      background: #d9d9d9;
-      height: 29px;
-      border-radius: 8px;
-      white-space: nowrap;
-    }
-  }
-  > .mapContainer {
-    padding: 0 16px;
-    margin-bottom: 28px;
-
-    > .map {
-      width: 100%;
-      height: 156px;
-    }
-  }
-}
-
-.textSection {
-  width: 100%;
-  border-bottom: 1px solid #8f8f8f;
-
-  > textarea {
-    border: none;
-    font: inherit;
-    font-size: 16px;
-    font-weight: 400;
-    width: 100%;
-    padding: 20px 16.5px;
-    resize: none;
-  }
-}
-
-.imageSection {
-  width: 100%;
-  display: flex;
-  gap: 5px;
-  padding: 16px 17.5px;
-  overflow-x: auto;
-  border-bottom: 1px solid #8f8f8f;
-
-  > .inputImage {
-    position: relative;
-    > label {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 101px;
-      height: 113px;
-      background-color: gray;
-    }
-    > input {
-      position: absolute;
-      width: 0;
-      height: 0;
-      padding: 0;
-    }
-  }
-
-  > .imagePreview {
-    height: 113px;
-    width: 101px;
-    object-fit: contain;
-  }
-}
-
-.voiceSection {
-  position: relative;
-  width: 100%;
-  display: flex;
-  padding: 20px 0;
-  border-bottom: 1px solid #8f8f8f;
-
-  > .recordButton {
-    width: 42px;
-    height: 42px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-left: 17px;
-    border-radius: 17px;
-    background-color: #d9d9d9;
-    border: none;
-
-    &.recording {
-      background-color: #369ed9;
-    }
-  }
-  > .playerWrapper {
-    width: calc(100% - 59px);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    > audio {
-      height: 42px;
-    }
-  }
-}
-
-.submitSection {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 30px 0;
-  > .mainPasteButton {
-    width: 85px;
-    height: 41px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font: inherit;
-    font-size: 16px;
-    font-weight: 700;
-    color: white;
-    border: none;
-    background-color: #d9d9d9;
-    &.active {
-      background-color: #369ed9;
-    }
-  }
-}
